Limit pod deletion to podman runtime
run pod deletion only if using podman, as Docker does not have a pod

+++ host_cleanup.sh
@@ -11,8 +11,10 @@ for name in ironic ironic-inspector dnsmasq httpd mariadb; do
 done
 
 # Remove existing pod
-if  sudo ${CONTAINER_RUNTIME} pod exists ironic-pod ; then
-    sudo ${CONTAINER_RUNTIME} pod rm ironic-pod -f
+if [[ "${CONTAINER_RUNTIME}" == "podman" ]]; then
+  if  sudo ${CONTAINER_RUNTIME} pod exists ironic-pod ; then
+      sudo ${CONTAINER_RUNTIME} pod rm ironic-pod -f
+  fi
 fi
 
 ANSIBLE_FORCE_COLOR=true ansible-playbook \
